KHWE-5: merge components in a one structure, host album images and delete the local ones
