--- src/components/Albums/AlbumCard/AlbumCard.scss
@@ -1,21 +1,0 @@
-.album-card {
-  margin: 0 0.9rem 2.5rem 0.9rem;
-  width: 22.5rem;
-
-  &__images {
-    width: 22.5rem;
-  }
-
-  &__description {
-    padding: 1rem;
-  }
-
-  &__card-album {
-    margin: 0;
-  }
-
-  &__card-artist {
-    color: rgba(0, 0, 0, 0.54);
-    margin: 0;
-  }
-}
