--- src/components/Albums/AlbumList/AlbumList.scss
@@ -1,14 +1,0 @@
-.album-list {
-  max-width: 100rem;
-  margin: 0 auto;
-
-  &__heading {
-    margin-left: 1.8rem;
-  }
-
-  &__container {
-    display: flex;
-    justify-content: flex-end;
-    flex-wrap: wrap;
-  }
-}
