+++ src/components/Albums/Albums.html
@@ -1,0 +1,26 @@
+<div class="albums">
+  <h2 class="albums__heading mdc-typography--headline4">Albums</h2>
+  <div class="albums__container">
+    <%  albums.forEach( album => {%>
+    <div class="albums__card-container">
+      <div class="albums__card mdc-card mdc-card-album">
+        <div class="albums__card-ripple-effect mdc-card__primary-action mdc-ripple-upgraded">
+          <div class="albums__card-media mdc-card__media mdc-card__media-first mdc-card__media--square">
+            <div class="mdc-card__media-content">
+              <img class="albums__card-images" src=<%=album.imageSource %>>
+            </div>
+          </div>
+          <div class="albums__card-description">
+            <h2 class="albums__card-album mdc-typography--headline6">
+              <%= album.albumName %>
+            </h2>
+            <h3 class="albums__card-artist mdc-typography--subtitle2">
+              <%= album.artistName %>
+            </h3>
+          </div>
+        </div>
+      </div>
+    </div>
+    <% }) %>
+  </div>
+</div>
