+++ src/components/Albums/Albums.scss
@@ -1,0 +1,34 @@
+.albums {
+  max-width: 100rem;
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+
+  &__container {
+    display: flex;
+    justify-content: flex-start;
+    flex-wrap: wrap;
+  }
+
+  &__card-container {
+    margin: 0 1.8rem 2.5rem 0;
+    width: 22.5rem;
+  }
+
+  &__card-images {
+    width: 22.5rem;
+  }
+
+  &__card-description {
+    padding: 1rem;
+  }
+
+  &__card-album {
+    margin: 0;
+  }
+
+  &__card-artist {
+    color: rgba(0, 0, 0, 0.54);
+    margin: 0;
+  }
+}
