+++ src/components/Main/Main.js
@@ -3,7 +3,7 @@ import { MDCDrawer } from "@material/drawer";
 import { PlayerComponent } from "../Player/Player";
 import { HeaderComponent } from "../Header/Header";
 import { SearchComponent } from "../Search/Search";
-import { AlbumList } from "../Albums/AlbumList/AlbumList";
+import { Albums } from "../Albums/Albums";
 import { AboutComponent } from "../About/About";
 import mainTemplate from "./Main.html";
 
@@ -68,7 +68,7 @@ export class MainComponent {
 
     this.about = new AboutComponent(this.mainPoint);
 
-    this.albums = new AlbumList(this.mainPoint);
+    this.albums = new Albums(this.mainPoint);
   }
 
   render() {
