+++ src/index.scss
@@ -16,7 +16,6 @@
 @import "./components/Main/Main";
 @import "./components/Player/Player";
 @import "./components/Page/Page";
-@import "components/Albums/AlbumList/AlbumList";
-@import "components/Albums/AlbumCard/AlbumCard";
+@import "components/Albums/Albums";
 @import "./components/About/About";
 @import "./components/DotsMenu/DotsMenu";
